requirements.txt
flask==2.3.2
flask-limiter==2.7.0
python-dotenv==1.0.0
requests==2.31.0